:recycle: refactor(api): move `entity-validator` to `src`
