+++ api/lib/domain/models/CompetenceMark.js
@@ -1,5 +1,5 @@
 import Joi from 'joi';
-import { validateEntity } from '../validators/entity-validator.js';
+import { validateEntity } from '../../../src/shared/domain/validators/entity-validator.js';
 
 const schema = Joi.object({
   id: Joi.number().integer().optional(),
