+++ api/src/shared/domain/models/AdminMember.js
@@ -1,7 +1,7 @@
 import BaseJoi from 'joi';
 import JoiDate from '@joi/date';
 const Joi = BaseJoi.extend(JoiDate);
-import { validateEntity } from '../../../../lib/domain/validators/entity-validator.js';
+import { validateEntity } from '../validators/entity-validator.js';
 import lodash from 'lodash';
 import { PIX_ADMIN } from '../../../authorization/domain/constants.js';
