+++ api/src/shared/domain/validators/entity-validator.js
@@ -1,4 +1,4 @@
-import { ObjectValidationError } from '../errors.js';
+import { ObjectValidationError } from '../../../../lib/domain/errors.js';
 
 const validateEntity = function (schema, entity) {
   const { error } = schema.validate(entity);
