+++ api/tests/unit/domain/validators/entity-validator_test.js
@@ -1,7 +1,7 @@
 import { expect } from '../../../test-helper.js';
 import { ObjectValidationError } from '../../../../lib/domain/errors.js';
 import Joi from 'joi';
-import { validateEntity } from '../../../../lib/domain/validators/entity-validator.js';
+import { validateEntity } from '../../../../src/shared/domain/validators/entity-validator.js';
 
 describe('Unit | Domain | Validators | entity-validator', function () {
   describe('#validateEntity', function () {
